requirements.txt
Flask
gunicorn
sklearn
pandas
numpy
PyYAML
evidently
dill
kaggle
import-ipynb
-e .